+++ vck-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/openid/PreRegisteredClientTest.kt
@@ -169,9 +169,10 @@ val PreRegisteredClientTest by matrixSuite {
         }
 
         "wrong client nonce in id_token should lead to error" {
+            val clientIdScheme = ClientIdScheme.PreRegistered(it.clientId, it.redirectUrl)
             it.verifierOid4vp = OpenId4VpVerifier(
                 keyMaterial = it.verifierKeyMaterial,
-                clientIdScheme = ClientIdScheme.PreRegistered(it.clientId, it.redirectUrl),
+                clientIdScheme = clientIdScheme,
                 nonceService = object : NonceService {
                     override suspend fun provideNonce() = uuid4().toString()
                     override suspend fun verifyNonce(it: String) = false